add linting & consistent formatting
I've run a format on all the files, and standardized file names / places to match the conventions I know of.

For VSCode, install Prettier, and ESLint extensions:
1. Go to settings
2. Use the prettier formatter
3. Set formatting on save
As for others editors: I'll help you set it up if it's problematic to install ESLint + Prettier, and format on save.

These are just my default settings and are up for change, BUT I'd *prefer* no semi colons or trailing commas (irrelevant to look at unless needed), and the widely used airbnb style-guide (comfort).